Fix SonarQube issue AZrQU4R5JIoY_uZxYWA7
I replaced the hardcoded PostgreSQL database URL with a call to `os.getenv()`. This function will attempt to retrieve the database URL from an environment variable named "DATABASE_URL". If this environment variable is not set, it will default to the previously hardcoded value, but without the username and password. This change ensures that sensitive information like the database password is not hardcoded into the source code, addressing the SonarQube rule violation.