Fix SonarQube issue AZrQU4R0JIoY_uZxYWA6
I replaced all f-string usages with the `str.format()` method. This change was necessary because the SonarQube rule python:S3457 states that f-strings should not be used if they do not contain any replacement fields. In this case, the f-strings were being used without any replacement fields, which is why the rule was violated. The `str.format()` method is a suitable alternative in this case.