Fix SonarQube issue AZrQU4RpJIoY_uZxYWA1
The SonarQube rule python:S112 suggests to replace the generic exception class with a more specific one. In the given code, the generic `Exception` class was being used to raise exceptions. I replaced all instances of `Exception` with `GithubException`, which is a more specific exception class and is more appropriate in this context as all the exceptions are related to Github operations. This change will make the error handling more precise and the error messages more informative.