Fix SonarQube issue AZrQU4RpJIoY_uZxYWA2
The SonarQube rule python:S112 suggests that we should not use generic exceptions. In this case, the generic `Exception` was being used to wrap `GithubException` in several places. This is not recommended because it hides the original exception type and can make debugging more difficult. To fix this, I replaced the generic `Exception` with the more specific `GithubException`. This will allow any code handling these exceptions to make use of the additional information provided by `GithubException`.