Fix SonarQube issue AZrQU4RpJIoY_uZxYWA4
The SonarQube rule python:S112 recommends replacing generic exceptions with more specific ones. In this case, the generic `Exception` class was being used to raise exceptions. I replaced these with the `GithubException` class, which is more specific and provides more context about the type of error that occurred. This change will make debugging easier and the code more robust.