Fix SonarQube issue AZrQU4ScJIoY_uZxYWBB
The SonarQube rule docker:S6476 requires the `as` keyword in Dockerfile to be in upper case format `AS`. This is because Dockerfile syntax is case-sensitive and the convention is to use uppercase for Dockerfile keywords. The only change made was replacing `as` with `AS` in the first line of the Dockerfile.