Fix SonarQube issue AZrQU4SjJIoY_uZxYWBF
I replaced the hardcoded PostgreSQL password with an environment variable `${POSTGRES_PASSWORD}`. This change was made to address the SonarQube rule violation that flagged the hardcoded password as a security risk. Now, the password can be set in the environment where this Docker Compose file is run, keeping it out of the codebase and enhancing security.