Fix SonarQube issue AZrVn4UOqVCJmKrPrIpz
The SonarQube rule python:S112 suggests that we should not use generic exceptions, and instead use specific exceptions. In this case, the code was raising a generic `Exception` type. I replaced all instances of `Exception` with `GithubException`, which is a more specific exception type provided by the Github API library. This change will make the exceptions more meaningful and easier to handle for the callers of these methods.